Adding enitities: transaction, and activity

+++ Fair2Share/Models/FriendRequests.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 
 namespace Fair2Share.Models {
     public class FriendRequests {
-        public int UserId { get; set; }
-        public int FutureFriendId { get; set; }
+        public long UserId { get; set; }
+        public long FutureFriendId { get; set; }
         public DateTime TimeStamp { get; set; }
         //public int State { get; set; }
     }
