+++ Fair2Share/Models/IProfileRepository.cs
@@ -1,6 +1,6 @@
 ﻿namespace Fair2Share.Models {
     public interface IProfileRepository {
-        Profile GetBy(int id);
+        Profile GetBy(long id);
         Profile GetBy(string email);
         void Add(Profile profile);
         void SaveChanges();
